Removed vanilla template support (BSA), updated hardcoded default TXT settings
Removed vanilla template support. Vanilla plugins have their facegen files inside BSA. The patcher can't access them, but it tried to "assume" that it doesn't need them. This was wrong as it resulted with black face or deformed face issuess.
Updated hardcoded TXT default settings that are used to create the TXT files if the patcher won't find them in SkyLady folder. Improved their instructions and simplified the default settings.